==> Temporal_Descriptives/00_queryingTrx.ipynb <==
## Trying to connect to the server via Python

import os
import sys
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

%matplotlib inline
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:80% !important; }</style>"))

import pandas as pd
import matplotlib
import matplotlib.pyplot as plt
import matplotlib.dates as mdates
import numpy as np
import datetime as dt
import time
import psycopg2
import psycopg2.extras

from Utils import TransantiagoConstants

DTPM_TRXDir = TransantiagoConstants.DTPM_TRXDir

* Connecting to the server and executing query for 2018

conn = psycopg2.connect(dbname='trxppu', user='lcamus', host='192.168.1.190', password='camus2018')

cursor = conn.cursor('cursor_unique_name', cursor_factory=psycopg2.extras.DictCursor)

query =  "SELECT unidad, patente, sersen, DATE(fhposini), sum(trx_validas), COUNT(*) FROM public.trxppu2018 GROUP BY unidad, patente, sersen, DATE(fhposini)"
cursor.execute(query)

file_2018 = os.path.join(DTPM_TRXDir,'un_ppu_sersen_sumtrx_2018_by_date.csv')
tic = time.clock()
with open(file_2018,"w") as f:
    for row in cursor:
        for element in row:        
            f.write('%s' % element + ';')
            
        f.write('\n')

f.close()
toc = time.clock()
print(toc-tic)

conn.close()

* Connecting to the server and executing query for 2017

conn = psycopg2.connect(dbname='trxppu', user='lcamus', host='192.168.1.190', password='camus2018')

cursor = conn.cursor('cursor_unique_name', cursor_factory=psycopg2.extras.DictCursor)

query =  "SELECT unidad, patente, sersen, DATE(fhposini), sum(trx_validas), COUNT(*) FROM public.trxppu2017 GROUP BY unidad, patente, sersen, DATE(fhposini)"
cursor.execute(query)

file_2017 = os.path.join(DTPM_TRXDir,'un_ppu_sersen_sumtrx_2017_by_date.csv')
tic = time.clock()
with open(file_2017,"w") as f:
    for row in cursor:
        for element in row:        
            f.write('%s' % element + ';')
            
        f.write('\n')

f.close()
toc = time.clock()
print(toc-tic)

conn.close()

* Connecting to the server and executing query for 2016

conn = psycopg2.connect(dbname='trxppu', user='lcamus', host='192.168.1.190', password='camus2018')

cursor = conn.cursor('cursor_unique_name', cursor_factory=psycopg2.extras.DictCursor)

query =  "SELECT unidad, patente, sersen, DATE(fhposini), sum(trx_validas), COUNT(*) FROM public.trxppu2016 GROUP BY unidad, patente, sersen, DATE(fhposini)"
cursor.execute(query)

file_2016 = os.path.join(DTPM_TRXDir,'un_ppu_sersen_sumtrx_2016_by_date.csv')
tic = time.clock()
with open(file_2016,"w") as f:
    for row in cursor:
        for element in row:        
            f.write('%s' % element + ';')
            
        f.write('\n')

f.close()
toc = time.clock()
print(toc-tic)

conn.close()

* Connecting to the server and executing query for 2015

conn = psycopg2.connect(dbname='trxppu', user='lcamus', host='192.168.1.190', password='camus2018')

cursor = conn.cursor('cursor_unique_name', cursor_factory=psycopg2.extras.DictCursor)

query =  "SELECT unidad, patente, sersen, DATE(fhposini), sum(trx_validas), COUNT(*) FROM public.trxppu2015 GROUP BY unidad, patente, sersen, DATE(fhposini)"
cursor.execute(query)

file_2015 = os.path.join(DTPM_TRXDir,'un_ppu_sersen_sumtrx_2015_by_date.csv')
tic = time.clock()
with open(file_2015,"w") as f:
    for row in cursor:
        for element in row:        
            f.write('%s' % element + ';')
            
        f.write('\n')

f.close()
toc = time.clock()
print(toc-tic)

conn.close()

### Closed
