==> Temporal_Descriptives/02_PrintingTimeSeriesToFiles.ipynb <==
# Printing grouped trx to files

import os
import sys
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

%matplotlib inline
#%matplotlib qt5
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:80% !important; }</style>"))

import pandas as pd
import matplotlib
import matplotlib.pyplot as plt
import matplotlib.dates as mdates
import numpy as np
import datetime as dt
import time

from Utils import TransantiagoConstants

DTPMDir = TransantiagoConstants.DTPMDir
DTPM_TRXDir = TransantiagoConstants.DTPM_TRXDir

from Utils import ReadTurnstilesDataBase
[ana_turnstiles_df, mauricio_turnstiles_df] = ReadTurnstilesDataBase.readTurnstileData()
ana_turnstiles_df = ReadTurnstilesDataBase.processAnaTurnstiles(ana_turnstiles_df)

import groupingSONDAData

* Grouping pn and zp by month

pn_output_path = os.path.join(DTPM_TRXDir,'1_MONTHLY/monthly_pn.csv')
zp_output_path = os.path.join(DTPM_TRXDir,'1_MONTHLY/monthly_zp.csv')

final_pn_grouped_data = groupingSONDAData.grouping_pn_data('by month', ana_turnstiles_df, mauricio_turnstiles_df)
final_zp_grouped_data = groupingSONDAData.grouping_zp_data('by month')

final_pn_grouped_data.reset_index(inplace=True)
final_pn_grouped_data.to_csv(pn_output_path,sep=';',encoding='latin-1')

final_zp_grouped_data.reset_index(inplace=True)
final_zp_grouped_data.to_csv(zp_output_path,sep=';',encoding='latin-1')

* Grouping pn and zp by month by un

pn_output_path = os.path.join(DTPM_TRXDir,'2_MONTHLY_UN/monthly_un_pn.csv')
zp_output_path = os.path.join(DTPM_TRXDir,'2_MONTHLY_UN/monthly_un_zp.csv')

final_pn_grouped_data = groupingSONDAData.grouping_pn_data('by month by un', ana_turnstiles_df, mauricio_turnstiles_df)
final_zp_grouped_data = groupingSONDAData.grouping_zp_data('by month by un')

final_pn_grouped_data.reset_index(inplace=True)
final_pn_grouped_data.to_csv(pn_output_path,sep=';',encoding='latin-1')

final_zp_grouped_data.reset_index(inplace=True)
final_zp_grouped_data.to_csv(zp_output_path,sep=';',encoding='latin-1')

* Grouping pn and zp by day

pn_output_path = os.path.join(DTPM_TRXDir,'3_DAILY/daily_pn.csv')
zp_output_path = os.path.join(DTPM_TRXDir,'3_DAILY/daily_zp.csv')

final_pn_grouped_data = groupingSONDAData.grouping_pn_data('by day', ana_turnstiles_df, mauricio_turnstiles_df)
final_zp_grouped_data = groupingSONDAData.grouping_zp_data('by day')

final_pn_grouped_data.reset_index(inplace=True)
final_pn_grouped_data.to_csv(pn_output_path,sep=';',encoding='latin-1')

final_zp_grouped_data.reset_index(inplace=True)
final_zp_grouped_data.to_csv(zp_output_path,sep=';',encoding='latin-1')

* Grouping pn and zp by day by un

pn_output_path = os.path.join(DTPM_TRXDir,'4_DAILY_UN/daily_un_pn.csv')
zp_output_path = os.path.join(DTPM_TRXDir,'4_DAILY_UN/daily_un_zp.csv')

final_pn_grouped_data = groupingSONDAData.grouping_pn_data('by day by un', ana_turnstiles_df, mauricio_turnstiles_df)
final_zp_grouped_data = groupingSONDAData.grouping_zp_data('by day by un')

final_pn_grouped_data.reset_index(inplace=True)
final_pn_grouped_data.to_csv(pn_output_path,sep=';',encoding='latin-1')

final_zp_grouped_data.reset_index(inplace=True)
final_zp_grouped_data.to_csv(zp_output_path,sep=';',encoding='latin-1')

## FINAL CLOSE
