==> Temporal_Descriptives/04_SummarizingTimesSeriesToFiles.ipynb <==
## Summarizing times series to file

import os
import sys
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

import summarizingSONDAData

summarizingSONDAData.reOrderingDataFrame('by month')

summarizingSONDAData.reOrderingDataFrame('by day')

summarizingSONDAData.reOrderingDataFrameByUN('by month by un')

summarizingSONDAData.reOrderingDataFrameByUN('by day by un')

## Final close
