==> Temporal_Descriptives/06_PlottingDynamicTimesSeries.ipynb <==
## Plotting times series - Interactive

import os
import sys
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

%matplotlib inline
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:80% !important; }</style>"))

import pandas as pd
import matplotlib
import matplotlib.pyplot as plt
import matplotlib.dates as mdates
from matplotlib.ticker import FuncFormatter
import numpy as np
import datetime as dt
import time

from Utils import TransantiagoConstants
DTPM_TRXDir = TransantiagoConstants.DTPM_TRXDir

import plotly.plotly
import plotly.graph_objs as go

* Defining function to plot

def plottingInteractively(dates,y_zp,y_st,y_3t,y_tm,y_exp, name, freq):
    trace1 = go.Bar(
        x=dates,
        y=y_zp,
        name='ZP',
        marker=dict(
            color='rgb(76, 181, 245)',
        ),
        opacity=0.6
    )

    trace2 = go.Bar(
        x=dates,
        y=y_st,
        name='S/T',
        marker=dict(
            color='rgb(183, 184, 182)',
        ),
        opacity=0.6
    )

    trace3 = go.Bar(
        x=dates,
        y=y_3t,
        name='C/T3',
        marker=dict(
            color='rgb(52, 103, 92)',
        ),
        opacity=0.6    

    )
    
    trace4 = go.Bar(
        x=dates,
        y=y_tm,
        name='C/TM',
        marker=dict(
            color='rgb(179, 193, 0)',
        ),
        opacity=0.6 
    )

    trace5 = go.Scatter(
        x = dates,
        y = y_exp,
        yaxis='y2',
        name = '%Exp.C/TM',
        marker=dict(
            color='rgb(0, 0, 0)',
        ),
    )


    data = [trace1, trace2, trace3, trace4, trace5]

    layout = go.Layout(
        title='Transacciones en bus y zona paga, ' + name + ', 2015 a 2017',
        barmode='stack',
        yaxis=dict(
            title='Transacciones'
        ),
        yaxis2=dict(
            title='% exps. diarias C/TM',
            overlaying='y',
            side='right',
            range=[0, 100]
        )
    )

    fig = go.Figure(data=data, layout=layout)
    plotly.offline.plot(fig, filename=name+'_'+freq)

## Plotting interactively

#### MENSUAL - SISTEMA

name = 'sistema'
freq = 'mensual'
monthly_input_path = os.path.join(DTPM_TRXDir,'1_MONTHLY/monthly_summary.csv')
monthly_trx = pd.read_csv(monthly_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01',end='2018-01',freq='M')
no_turnstile_values = list(map(int, monthly_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,monthly_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,monthly_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### MENSUAL - U1

name = 'U1'
freq = 'mensual'
monthly_input_path = os.path.join(DTPM_TRXDir,'2_MONTHLY_UN/' + name + '/' + name + '_summary.csv')
monthly_trx = pd.read_csv(monthly_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01',end='2018-01',freq='M')
no_turnstile_values = list(map(int, monthly_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,monthly_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,monthly_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### MENSUAL - U2

name = 'U2'
freq = 'mensual'
monthly_input_path = os.path.join(DTPM_TRXDir,'2_MONTHLY_UN/' + name + '/' + name + '_summary.csv')
monthly_trx = pd.read_csv(monthly_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01',end='2018-01',freq='M')
no_turnstile_values = list(map(int, monthly_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,monthly_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,monthly_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### MENSUAL - U3

name = 'U3'
freq = 'mensual'
monthly_input_path = os.path.join(DTPM_TRXDir,'2_MONTHLY_UN/' + name + '/' + name + '_summary.csv')
monthly_trx = pd.read_csv(monthly_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01',end='2018-01',freq='M')
no_turnstile_values = list(map(int, monthly_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,monthly_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,monthly_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### MENSUAL - U4

name = 'U4'
freq = 'mensual'
monthly_input_path = os.path.join(DTPM_TRXDir,'2_MONTHLY_UN/' + name + '/' + name + '_summary.csv')
monthly_trx = pd.read_csv(monthly_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01',end='2018-01',freq='M')
no_turnstile_values = list(map(int, monthly_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,monthly_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,monthly_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### MENSUAL - U5

name = 'U5'
freq = 'mensual'
monthly_input_path = os.path.join(DTPM_TRXDir,'2_MONTHLY_UN/' + name + '/' + name + '_summary.csv')
monthly_trx = pd.read_csv(monthly_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01',end='2018-01',freq='M')
no_turnstile_values = list(map(int, monthly_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,monthly_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,monthly_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### MENSUAL - U6

name = 'U6'
freq = 'mensual'
monthly_input_path = os.path.join(DTPM_TRXDir,'2_MONTHLY_UN/' + name + '/' + name + '_summary.csv')
monthly_trx = pd.read_csv(monthly_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01',end='2018-01',freq='M')
no_turnstile_values = list(map(int, monthly_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,monthly_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,monthly_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### MENSUAL - U7

name = 'U7'
freq = 'mensual'
monthly_input_path = os.path.join(DTPM_TRXDir,'2_MONTHLY_UN/' + name + '/' + name + '_summary.csv')
monthly_trx = pd.read_csv(monthly_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01',end='2018-01',freq='M')
no_turnstile_values = list(map(int, monthly_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,monthly_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,monthly_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,monthly_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### DIARIO - SISTEMA

name = 'sistema'
freq = 'diario'
daily_input_path = os.path.join(DTPM_TRXDir,'3_DAILY/daily_summary.csv')
daily_trx = pd.read_csv(daily_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01-01',end='2017-12-31',freq='D')
no_turnstile_values = list(map(int, daily_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,daily_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,daily_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### DIARIO - U1

name = 'U1'
freq = 'diario'
daily_input_path = os.path.join(DTPM_TRXDir,'4_DAILY_UN/' + name + '/' + name + '_summary.csv')
daily_trx = pd.read_csv(daily_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01-01',end='2017-12-31',freq='D')
no_turnstile_values = list(map(int, daily_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,daily_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,daily_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### DIARIO - U2

name = 'U2'
freq = 'diario'
daily_input_path = os.path.join(DTPM_TRXDir,'4_DAILY_UN/' + name + '/' + name + '_summary.csv')
daily_trx = pd.read_csv(daily_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01-01',end='2017-12-31',freq='D')
no_turnstile_values = list(map(int, daily_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,daily_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,daily_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### DIARIO - U3

name = 'U3'
freq = 'diario'
daily_input_path = os.path.join(DTPM_TRXDir,'4_DAILY_UN/' + name + '/' + name + '_summary.csv')
daily_trx = pd.read_csv(daily_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01-01',end='2017-12-31',freq='D')
no_turnstile_values = list(map(int, daily_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,daily_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,daily_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### DIARIO - U4

name = 'U4'
freq = 'diario'
daily_input_path = os.path.join(DTPM_TRXDir,'4_DAILY_UN/' + name + '/' + name + '_summary.csv')
daily_trx = pd.read_csv(daily_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01-01',end='2017-12-31',freq='D')
no_turnstile_values = list(map(int, daily_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,daily_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,daily_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### DIARIO - U5

name = 'U5'
freq = 'diario'
daily_input_path = os.path.join(DTPM_TRXDir,'4_DAILY_UN/' + name + '/' + name + '_summary.csv')
daily_trx = pd.read_csv(daily_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01-01',end='2017-12-31',freq='D')
no_turnstile_values = list(map(int, daily_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,daily_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,daily_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### DIARIO - U6

name = 'U6'
freq = 'diario'
daily_input_path = os.path.join(DTPM_TRXDir,'4_DAILY_UN/' + name + '/' + name + '_summary.csv')
daily_trx = pd.read_csv(daily_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01-01',end='2017-12-31',freq='D')
no_turnstile_values = list(map(int, daily_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,daily_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,daily_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

#### DIARIO - U7

name = 'U7'
freq = 'diario'
daily_input_path = os.path.join(DTPM_TRXDir,'4_DAILY_UN/' + name + '/' + name + '_summary.csv')
daily_trx = pd.read_csv(daily_input_path,sep=';',encoding='latin-1', index_col=0)

dates = pd.date_range(start='2015-01-01',end='2017-12-31',freq='D')
no_turnstile_values = list(map(int, daily_trx['pn_SUM_TRX_no_t'].values))
three_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_3t'].values))
butterfly_turnstile_values = list(map(int,daily_trx['pn_SUM_TRX_tm'].values))
zp_values = list(map(int,daily_trx['zp_SUM_TRX'].values))
ratio_butterfly_turnstile_values = list(map(int,daily_trx['ratio_tm'].values))

plottingInteractively(dates,zp_values,no_turnstile_values,three_turnstile_values,butterfly_turnstile_values,ratio_butterfly_turnstile_values,name,freq)

## Closed
